requirements.txt
openai>=1.0.0
requests>=2.31.0
feedparser>=6.0.11
python-dateutil>=2.9.0
notion-client>=2.2.1